refactor: complete migration to multi-layered architecture, cleanup src and update test suite

+++ src/DTO/AccumulatorState.php
@@ -1,34 +1,36 @@
-<?php
-
-declare(strict_types=1);
+<?php declare(strict_types=1);
 
 namespace Aleoosha\HiveMind\DTO;
 
 /**
- * Внутренний объект для накопления итогов без хранения истории всех точек.
+ * Internal state for aggregating metrics before archiving.
+ * Stores values as integers (FixedPoint raw values) to maintain precision.
  */
 final class AccumulatorState
 {
-    public float $sumHealth = 0.0;
-    public float $sumCpu = 0.0;
-    public float $maxCpu = 0.0;
-    public float $sumDb = 0.0;
-    public float $maxDb = 0.0;
-    public float $sumApi = 0.0;
-    public float $maxApi = 0.0;
-    public float $sumShedding = 0.0;
+    public int $sumHealth = 0;
+    public int $sumCpu = 0;
+    public int $maxCpu = 0;
+    public int $sumDb = 0;
+    public int $maxDb = 0;
+    public int $sumApi = 0;
+    public int $maxApi = 0;
+    public int $sumShedding = 0;
     public int $count = 0;
 
+    /**
+     * Reset all counters to zero for the next accumulation window.
+     */
     public function reset(): void
     {
-        $this->sumHealth = 0.0;
-        $this->sumCpu = 0.0;
-        $this->maxCpu = 0.0;
-        $this->sumDb = 0.0;
-        $this->maxDb = 0.0;
-        $this->sumApi = 0.0;
-        $this->maxApi = 0.0;
-        $this->sumShedding = 0.0;
+        $this->sumHealth = 0;
+        $this->sumCpu = 0;
+        $this->maxCpu = 0;
+        $this->sumDb = 0;
+        $this->maxDb = 0;
+        $this->sumApi = 0;
+        $this->maxApi = 0;
+        $this->sumShedding = 0;
         $this->count = 0;
     }
 }
